Styling home page.
Changed layout from one that was responsive to one that is fixed and
centered on a background. I did this because it is a site that will
most likely not be viewed on a mobile device. Also, since the content
is so sparse and text-heavy, the background image adds needed texture
and interest.